migrated torrents api resource from category to torrent file
